<==== Mercurial data ====>
Branch: 3.1_singletone_refactor

+++ common/src/network/h264_rtp_parser.cpp
@@ -170,7 +170,7 @@ bool CLH264RtpParser::isBufferOverflow() const
         addHeaderSize = getSpsPpsSize();
 
     int totalSize = m_videoFrameSize + addHeaderSize;
-    return totalSize > MAX_ALLOWED_FRAME_SIZE;
+    return totalSize > (int) MAX_ALLOWED_FRAME_SIZE;
 }
 
 QnCompressedVideoDataPtr CLH264RtpParser::createVideoData(
@@ -439,7 +439,7 @@ bool CLH264RtpParser::processData(
     bool isPacketLost = m_prevSequenceNum != -1
         && quint16(m_prevSequenceNum) != quint16(sequenceNum-1);
 
-    if (m_videoFrameSize > MAX_ALLOWED_FRAME_SIZE)
+    if (m_videoFrameSize > (int) MAX_ALLOWED_FRAME_SIZE)
     {
         NX_LOG("Too large RTP/H.264 frame. Truncate video buffer", cl_logWARNING);
         clearInternalBuffer();
